--- src/main/java/org/crops/fitserver/auth/facade/dto/SocialLoginPageResponse.java
@@ -1,16 +1,0 @@
-package org.crops.fitserver.auth.facade.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class SocialLoginPageResponse {
-
-  private String loginPageUrl;
-
-  public static SocialLoginPageResponse from(String loginPageUrl) {
-    return new SocialLoginPageResponse(loginPageUrl);
-  }
-}
