--- src/main/java/org/crops/fitserver/auth/service/AuthService.java
@@ -1,5 +1,0 @@
-package org.crops.fitserver.auth.service;
-
-public interface AuthService {
-
-}
