--- src/main/java/org/crops/fitserver/auth/service/OAuthService.java
@@ -1,16 +1,0 @@
-package org.crops.fitserver.auth.service;
-
-
-import org.crops.fitserver.user.domain.SocialPlatform;
-import org.crops.fitserver.user.domain.User;
-
-public interface OAuthService {
-
-  boolean support(SocialPlatform socialPlatform);
-
-  User socialUserLogin(
-      String redirectUrl,
-      String authorizationCode);
-
-  String getLoginPageUrl();
-}
