--- src/main/java/org/crops/fitserver/core/file/constant/FileDomain.java
@@ -1,15 +1,0 @@
-package org.crops.fitserver.core.file.constant;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum FileDomain {
-  PROFILE_IMAGE("/profile", false),
-  PORTFOLIO("/portfolio", false),
-  CHAT("/chat", true);
-
-  private final String directory;
-  private final boolean isTemporary;
-}
