--- src/main/java/org/crops/fitserver/core/file/dto/PreSignedUrlDto.java
@@ -1,11 +1,0 @@
-package org.crops.fitserver.core.file.dto;
-
-import lombok.Builder;
-
-@Builder
-public record PreSignedUrlDto(
-    String preSignedUrl,
-    String fileKey
-) {
-
-}
