+++ src/main/java/org/crops/fitserver/domain/auth/facade/dto/SocialLoginPageResponse.java
@@ -1,0 +1,16 @@
+package org.crops.fitserver.domain.auth.facade.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class SocialLoginPageResponse {
+
+  private String loginPageUrl;
+
+  public static SocialLoginPageResponse from(String loginPageUrl) {
+    return new SocialLoginPageResponse(loginPageUrl);
+  }
+}
