+++ src/main/java/org/crops/fitserver/domain/auth/facade/dto/TokenResponse.java
@@ -1,0 +1,20 @@
+package org.crops.fitserver.domain.auth.facade.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.crops.fitserver.global.jwt.TokenCollection;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class TokenResponse {
+
+  private final String accessToken;
+  private final String refreshToken;
+
+  public static TokenResponse of(TokenCollection tokenCollection) {
+    return new TokenResponse(
+        tokenCollection.getAccessToken(),
+        tokenCollection.getRefreshToken());
+  }
+}
