+++ src/main/java/org/crops/fitserver/domain/auth/service/AuthService.java
@@ -1,0 +1,5 @@
+package org.crops.fitserver.domain.auth.service;
+
+public interface AuthService {
+
+}
