+++ src/main/java/org/crops/fitserver/domain/auth/service/OAuthService.java
@@ -1,0 +1,16 @@
+package org.crops.fitserver.domain.auth.service;
+
+
+import org.crops.fitserver.domain.user.domain.SocialPlatform;
+import org.crops.fitserver.domain.user.domain.User;
+
+public interface OAuthService {
+
+  boolean support(SocialPlatform socialPlatform);
+
+  User socialUserLogin(
+      String redirectUrl,
+      String authorizationCode);
+
+  String getLoginPageUrl();
+}
