+++ src/main/java/org/crops/fitserver/domain/file/dto/PreSignedUrlDto.java
@@ -1,0 +1,11 @@
+package org.crops.fitserver.domain.file.dto;
+
+import lombok.Builder;
+
+@Builder
+public record PreSignedUrlDto(
+    String preSignedUrl,
+    String fileKey
+) {
+
+}
