+++ src/main/java/org/crops/fitserver/domain/user/domain/UserRole.java
@@ -1,0 +1,9 @@
+package org.crops.fitserver.domain.user.domain;
+
+public enum UserRole {
+
+  NON_MEMBER,
+  MEMBER,
+  MANAGER,
+  ADMIN;
+}
