+++ src/main/java/org/crops/fitserver/global/http/SocailPlatformRequestConverter.java
@@ -1,6 +1,6 @@
 package org.crops.fitserver.global.http;
 
-import org.crops.fitserver.user.domain.SocialPlatform;
+import org.crops.fitserver.domain.user.domain.SocialPlatform;
 import org.springframework.core.convert.converter.Converter;
 
 public class SocailPlatformRequestConverter implements Converter<String, SocialPlatform> {
