+++ src/main/java/org/crops/fitserver/global/security/PrincipalDetails.java
@@ -5,7 +5,7 @@ import java.util.Collection;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
-import org.crops.fitserver.user.domain.User;
+import org.crops.fitserver.domain.user.domain.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
