--- src/main/java/org/crops/fitserver/user/domain/UserRole.java
@@ -1,9 +1,0 @@
-package org.crops.fitserver.user.domain;
-
-public enum UserRole {
-
-  NON_MEMBER,
-  MEMBER,
-  MANAGER,
-  ADMIN;
-}
