--- src/main/java/org/crops/fitserver/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package org.crops.fitserver.user.repository;
-
-import java.util.Optional;
-import org.crops.fitserver.user.domain.User;
-import org.springframework.data.repository.Repository;
-
-public interface UserRepository extends
-    Repository<User, Long> {
-
-  Optional<User> findById(Long id);
-
-  User save(User user);
-}
